# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Launchpad Control Plane API
  description: |
    Self-hosted deployment control plane. Workspace isolation via bearer tokens.
    Ambient environment via `X-Launchpad-Environment` (default `dev`).
    Errors use RFC 7807 `application/problem+json` with optional `code` and `hints`.
  version: 0.1.0
servers:
  - url: http://localhost:8080
    description: Local default
security:
  - bearerAuth: []
...
paths:
...
  /v1/tokens:
    post:
      tags: [tokens]
      summary: Create API token (mints service account principal)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                workspace: { type: string, default: default }
                scopes:
                  type: array
                  items: { type: string }
      responses:
        "201":
          description: Token created (plaintext shown once)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenCreate"
        "401":
          $ref: "#/components/responses/Problem"
  /v1/audit:
    get:
      tags: [audit]
      summary: List recent workspace audit events
      parameters:
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Audit events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditEvent"
  /v1/jobs/{id}:
    get:
      tags: [jobs]
      summary: Get job by ID
      parameters:
        - $ref: "#/components/parameters/JobID"
      responses:
        "200":
          description: Job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          $ref: "#/components/responses/Problem"
  /v1/projects:
    get:
      tags: [projects]
      summary: List projects
      responses:
        "200":
          description: Projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      tags: [projects]
      summary: Create project (bootstrap dev env + primary service)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProject"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /v1/projects/{project}:
    get:
      tags: [projects]
      summary: Get project
      parameters:
        - $ref: "#/components/parameters/ProjectName"
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /v1/projects/{project}/config:
    get:
      tags: [config]
      summary: Get config (resolved or layer); secrets redacted as "***"
      parameters:
        - $ref: "#/components/parameters/ProjectName"
        - $ref: "#/components/parameters/Environment"
        - name: layer
          in: query
          schema:
            type: string
            enum: [shared, service, resolved]
        - name: view
          in: query
          description: "typed returns {value,sensitivity,set} per key; secret values are null"
          schema:
            type: string
            enum: [typed]
      responses:
        "200":
          description: Config map (secret values are "***") or typed entries when view=typed
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
    patch:
      tags: [config]
      summary: Patch service-layer config vars (sticky sensitivity; response redacted)
      parameters:
        - $ref: "#/components/parameters/ProjectName"
        - $ref: "#/components/parameters/Environment"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: string
                nullable: true
      responses:
        "200":
          description: Resolved config after patch (secrets as "***")
  /v1/targets/{type}/capabilities:
    get:
      tags: [runtime]
      summary: Target capabilities and extension field list
      parameters:
        - name: type
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Capabilities
  /v1/projects/{project}/processes:
    get:
      tags: [projects]
      summary: List process definitions
      parameters:
        - $ref: "#/components/parameters/ProjectName"
      responses:
        "200":
          description: Processes
  /v1/projects/{project}/logs:
    get:
      tags: [runtime]
      summary: Process logs (target-backed)
      parameters:
        - $ref: "#/components/parameters/ProjectName"
        - $ref: "#/components/parameters/Environment"
        - name: process
          in: query
          schema: { type: string, default: web }
      responses:
        "200":
          description: Plain text log stream/body
  /v1/projects/{project}/environments:
    get:
      tags: [environments]
      summary: List environments
      parameters:
        - $ref: "#/components/parameters/ProjectName"
      responses:
        "200":
          description: Environments
    post:
      tags: [environments]
      summary: Create environment
      parameters:
        - $ref: "#/components/parameters/ProjectName"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEnvironment"
      responses:
        "201":
          description: Created
  /v1/projects/{project}/environments/{name}:
    get:
      tags: [environments]
      summary: Get environment
      parameters:
        - $ref: "#/components/parameters/ProjectName"
        - name: name
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Environment
  /v1/projects/{project}/environments/{name}/clone:
    post:
      tags: [environments]
      summary: Clone environment (plain config copy; secrets as needs_value only)
      parameters:
        - $ref: "#/components/parameters/ProjectName"
        - name: name
          in: path
          required: true
          schema: { type: string }
          description: Source environment name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, description: Destination environment name }
                target:
                  type: object
                  properties:
                    type: { type: string }
                    namespace: { type: string }
                    cluster: { type: string }
                ephemeral: { type: boolean }
      responses:
        "201":
          description: Cloned environment + report
        "409":
          description: Destination already exists
  /v1/projects/{project}/releases:
    get:
      tags: [releases]
      summary: List releases (with deployment annotations)
      parameters:
        - $ref: "#/components/parameters/ProjectName"
        - $ref: "#/components/parameters/Environment"
      responses:
        "200":
          description: Releases
    post:
      tags: [releases]
      summary: Create image release and deploy
      parameters:
        - $ref: "#/components/parameters/ProjectName"
        - $ref: "#/components/parameters/Environment"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRelease"
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReleaseJob"
        "409":
          $ref: "#/components/responses/Problem"
  /v1/projects/{project}/rollback:
    post:
      tags: [releases]
      summary: Rollback to prior release version
      parameters:
        - $ref: "#/components/parameters/ProjectName"
        - $ref: "#/components/parameters/Environment"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [version]
              properties:
                version: { type: integer, minimum: 1 }
                description: { type: string }
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReleaseJob"
  /v1/projects/{project}/promote:
    post:
      tags: [releases]
      summary: Promote release across environments (re-resolve target config)
      parameters:
        - $ref: "#/components/parameters/ProjectName"
        - $ref: "#/components/parameters/Environment"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [from]
              properties:
                from: { type: string }
                to: { type: string }
                version: { type: integer }
                description: { type: string }
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReleaseJob"
  /v1/projects/{project}/changeset:
    get:
      tags: [changeset]
      summary: Get open changeset
      parameters:
        - $ref: "#/components/parameters/ProjectName"
        - $ref: "#/components/parameters/Environment"
      responses:
        "200":
          description: Changeset
    delete:
      tags: [changeset]
      summary: Discard open changeset
      parameters:
        - $ref: "#/components/parameters/ProjectName"
      responses:
        "204":
          description: Discarded
  /v1/projects/{project}/changeset/changes:
    post:
      tags: [changeset]
      summary: Stage changes
      parameters:
        - $ref: "#/components/parameters/ProjectName"
        - $ref: "#/components/parameters/Environment"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                changes:
                  type: array
                  items:
                    type: object
      responses:
        "200":
          description: Updated changeset
        "409":
          $ref: "#/components/responses/Problem"
  /v1/projects/{project}/changeset/changes/last:
    delete:
      tags: [changeset]
      summary: Unstage the most recently staged change
      parameters:
        - $ref: "#/components/parameters/ProjectName"
      responses:
        "200":
          description: Deleted change and remaining count
          content:
            application/json:
              schema:
                type: object
                properties:
                  change: { type: object }
                  remaining_count: { type: integer }
                  environment: { type: string }
        "404":
          $ref: "#/components/responses/Problem"
  /v1/projects/{project}/changeset/push:
    post:
      tags: [changeset]
      summary: Push staged changeset (materialize + deploy)
      parameters:
        - $ref: "#/components/parameters/ProjectName"
        - $ref: "#/components/parameters/Environment"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                description: { type: string }
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReleaseJob"
  /v1/projects/{project}/preview:
    get:
      tags: [changeset]
      summary: Pending vs baseline, release↔release, or env↔env structured preview
      parameters:
        - $ref: "#/components/parameters/ProjectName"
        - $ref: "#/components/parameters/Environment"
        - name: from_release
          in: query
          schema: { type: integer }
        - name: to_release
          in: query
          schema: { type: integer }
        - name: from_env
          in: query
          schema: { type: string }
          description: Baseline environment name (requires to_env; exclusive with release pair)
        - name: to_env
          in: query
          schema: { type: string }
          description: Target environment name (requires from_env; exclusive with release pair)
      responses:
        "200":
          description: Preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Preview"
  /v1/projects/{project}/manifest:
    get:
      tags: [projects]
      summary: Export live project manifest (launchpad.yaml v1 as JSON)
      description: |
        Returns live control-plane state, not the open changeset.
        Does not use X-Launchpad-Environment; filter with query `environment`.
      parameters:
        - $ref: "#/components/parameters/ProjectName"
        - name: environment
          in: query
          schema: { type: string }
          description: If set, environments map contains only this env (404 if missing)
      responses:
        "200":
          description: Manifest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Manifest"
        "404":
          $ref: "#/components/responses/Problem"
  /v1/projects/{project}/manifest/apply:
    post:
      tags: [projects]
      summary: Apply a v1 manifest (create-if-missing, stage only)
      parameters:
        - $ref: "#/components/parameters/ProjectName"
        - $ref: "#/components/parameters/Environment"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [document]
              properties:
                document:
                  $ref: "#/components/schemas/Manifest"
                environment:
                  type: string
                  description: Selected env; default is X-Launchpad-Environment or dev
      responses:
        "200":
          description: Apply report (never deploys)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplyReport"
        "400":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Workspace API token (`lp_…`) or bootstrap token
  parameters:
    ProjectName:
      name: project
      in: path
      required: true
      schema: { type: string }
    JobID:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, default: 50 }
  responses:
    Problem:
      description: RFC 7807 problem+json
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    Project:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        primary_service: { type: string }
        status: { type: string }
        created_at: { type: string, format: date-time }
    CreateProject:
      type: object
      required: [name]
      properties:
        name: { type: string }
        target:
          type: object
          properties:
            type: { type: string }
            namespace: { type: string }
            cluster: { type: string }
...
    Job:
      type: object
      properties:
        id: { type: string }
        type: { type: string }
        status: { type: string }
        last_error: { type: string }
    TokenCreate:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        workspace: { type: string }
        scopes:
          type: array
          items: { type: string }
        token: { type: string }
        principal_id: { type: string }
        principal_kind: { type: string }
    AuditEvent:
      type: object
      properties:
        id: { type: string }
        action: { type: string }
        resource_type: { type: string }
        resource_id: { type: string }
        principal_id: { type: string }
        token_id: { type: string }
        project_name: { type: string }
        detail:
          type: object
          additionalProperties: { type: string }
        created_at: { type: string, format: date-time }
...